Mudança no comportamento do web scraping no pastebin

+++ app/src/main/java/com/droidev/levantamento/Pastebin.java
@@ -5,6 +5,7 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.text.InputType;
+import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.LinearLayout;
@@ -196,11 +197,16 @@ public class Pastebin {
         devKey.setHint("Sua chave dev api");
         devKey.setInputType(InputType.TYPE_CLASS_TEXT);
 
+        EditText elemento = new EditText(context);
+        elemento.setHint("Elemento da pagina");
+        elemento.setInputType(InputType.TYPE_CLASS_TEXT);
+
         LinearLayout lay = new LinearLayout(context);
         lay.setOrientation(LinearLayout.VERTICAL);
         lay.addView(login);
         lay.addView(senha);
         lay.addView(devKey);
+        lay.addView(elemento);
 
         AlertDialog dialog = new AlertDialog.Builder(context)
                 .setCancelable(false)
@@ -217,41 +223,51 @@ public class Pastebin {
         login.setText(tinyDB.getString("login"));
         senha.setText(tinyDB.getString("senha"));
         devKey.setText(tinyDB.getString("devKey"));
+        elemento.setText(tinyDB.getString("elemento"));
 
         Button positiveButton = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
         Button neutralButton = dialog.getButton(AlertDialog.BUTTON_NEUTRAL);
 
-        positiveButton.setOnClickListener(v -> {
+        positiveButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
 
-            String loginString = login.getText().toString();
-            String senhaString = senha.getText().toString();
-            String devKeyString = devKey.getText().toString();
+                String loginString = login.getText().toString();
+                String senhaString = senha.getText().toString();
+                String devKeyString = devKey.getText().toString();
+                String elementoString = elemento.getText().toString();
 
-            if (!loginString.equals("") && !senhaString.equals("") && !devKeyString.equals("")) {
+                if (!loginString.equals("") && !senhaString.equals("") && !devKeyString.equals("") && !elementoString.equals("")) {
 
-                tinyDB.remove("login");
-                tinyDB.remove("senha");
-                tinyDB.remove("devKey");
+                    tinyDB.remove("login");
+                    tinyDB.remove("senha");
+                    tinyDB.remove("devKey");
+                    tinyDB.remove("elemento");
 
-                tinyDB.putString("login", loginString);
-                tinyDB.putString("senha", senhaString);
-                tinyDB.putString("devKey", devKeyString);
+                    tinyDB.putString("login", loginString);
+                    tinyDB.putString("senha", senhaString);
+                    tinyDB.putString("devKey", devKeyString);
+                    tinyDB.putString("elemento", elementoString);
 
-                Toast.makeText(context, "Conta salva", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(context, "Conta salva", Toast.LENGTH_SHORT).show();
 
-                dialog.dismiss();
+                    dialog.dismiss();
 
-            } else {
+                } else {
 
-                Toast.makeText(context, "Erro, campo vazio", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(context, "Erro, campo vazio", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
-        neutralButton.setOnClickListener(v -> {
+        neutralButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
 
-            login.setText("");
-            senha.setText("");
-            devKey.setText("");
+                login.setText("");
+                senha.setText("");
+                devKey.setText("");
+            }
         });
     }
 
@@ -267,9 +283,11 @@ public class Pastebin {
 
             try {
 
+                TinyDB tinyDB = new TinyDB(activity.getBaseContext());
+
                 Document doc = Jsoup.connect(url).get();
 
-                String text = doc.select("textarea[class=textarea]").text();
+                String text = doc.select(tinyDB.getString("elemento")).text();
 
                 sb.append(text);
